Expand and rework find_duplicate_users script

+++ tools/data_fixers/find_duplicate_users.py
@@ -1,13 +1,16 @@
-from collections import Counter
+from typing import Dict
+from typing import List
 from operator import itemgetter
 
 import click
 import requests
+from more_itertools import bucket
 from os2mo_helpers.mora_helpers import MoraHelper
 from ra_utils.load_settings import load_settings
+from ra_utils.apply import apply
 
 
-def check_duplicate_cpr(mora_base: str) -> list:
+def check_duplicate_cpr(mora_base: str) -> Dict[str, List[str]]:
 
     helper = MoraHelper(hostname=mora_base)
     users = helper.read_all_users()
@@ -27,7 +30,31 @@ def check_duplicate_cpr(mora_base: str) -> list:
     required=False,
     help="Remove all user objects that has the same cpr-number as another user",
 )
-def cli(delete):
+@click.option(
+    "--keep-one",
+    type=click.BOOL,
+    default=False,
+    is_flag=True,
+    required=False,
+    help="Instead of removing all, remove all, but one per duplicate set",
+)
+@click.option(
+    "--dry-run",
+    type=click.BOOL,
+    default=False,
+    is_flag=True,
+    required=False,
+    help="Write out all duplicates",
+)
+@click.option(
+    "--verbose",
+    type=click.BOOL,
+    default=False,
+    is_flag=True,
+    required=False,
+    help="Write out all duplicates",
+)
+def cli(delete: bool, keep_one: bool, dry_run: bool, verbose: bool):
     """Find users in MO that have the same CPR number.
 
     Prints the number of cpr-numbers that are used by more than one user and the list of uuids for the users sharing a cpr-number.
@@ -36,18 +63,34 @@ def cli(delete):
     settings = load_settings()
     mox_base = settings.get("mox.base")
     mora_base = settings.get("mora.base")
-    uuids = check_duplicate_cpr(mora_base)
+    duplicate_dict = check_duplicate_cpr(mora_base)
+
+    if verbose:
+        for cpr_no, uuids in duplicate_dict.items():
+            click.echo(f"CPR: {cpr_no}, with UUIDS:")
+            for uuid in uuids:
+                click.echo(f"\t{uuid}")
 
     if delete:
-        for uuid in uuids:
-            r = requests.delete(f"{mox_base}/organisation/bruger/{uuid}")
-            r.raise_for_status()
+        for cpr_no, uuids in duplicate_dict.items():
+            # If we are keeping one, forward iterator once
+            uuids = iter(uuids)
+            if keep_one:
+                next(uuids)
+            for uuid in uuids:
+                if dry_run:
+                    click.echo(f"Would have deleted {uuid}")
+                else:
+                    if verbose:
+                        click.echo(f"Deleting {uuid}")
+                    r = requests.delete(f"{mox_base}/organisation/bruger/{uuid}")
+                    r.raise_for_status()
     else:
+        if keep_one:
+            raise click.ClickException("Cannot pass 'keep_one' without 'delete'")
         click.echo(
-            f"There are {len(uuids)} CPR-number(s) assigned to more than one user"
+            f"There are {len(duplicate_dict)} CPR-number(s) assigned to more than one user"
         )
-        if uuids:
-            click.echo(uuids)
 
 
 if __name__ == "__main__":
